Refactors the SchemaBrowser to remove duplicate code and logic.
