+++ packages/front-end/components/SchemaBrowser/RetryInformationSchemaCard.tsx
@@ -1,21 +1,18 @@
-import { useEffect, useState } from "react";
+import { useState } from "react";
 import { InformationSchemaInterface } from "@/../back-end/src/types/Integration";
 import { useAuth } from "@/services/auth";
-import LoadingSpinner from "../LoadingSpinner";
 
 export default function RetryInformationSchemaCard({
   datasourceId,
-  mutate,
+  setFetching,
   informationSchema,
 }: {
   datasourceId: string;
-  mutate: () => void;
+  setFetching: (fetching: boolean) => void;
   informationSchema: InformationSchemaInterface;
 }) {
   const [error, setError] = useState<null | string>(null);
   const { apiCall } = useAuth();
-  const [fetching, setFetching] = useState(false);
-  const [retryCount, setRetryCount] = useState(1);
 
   async function onClick() {
     setError(null);
@@ -35,55 +32,20 @@ export default function RetryInformationSchemaCard({
     }
   }
 
-  useEffect(() => {
-    if (fetching) {
-      if (
-        retryCount > 1 &&
-        retryCount < 8 &&
-        informationSchema.status === "COMPLETE"
-      ) {
-        setFetching(false);
-        setRetryCount(1);
-      } else if (retryCount > 8) {
-        setFetching(false);
-        setError(
-          "This query is taking quite a while. We're building this in the background. Feel free to leave this page and check back in a few minutes."
-        );
-        setRetryCount(1);
-      } else {
-        const timer = setTimeout(() => {
-          mutate();
-          setRetryCount(retryCount * 2);
-        }, retryCount * 1000);
-        return () => {
-          clearTimeout(timer);
-        };
-      }
-    }
-  }, [fetching, mutate, retryCount, informationSchema]);
-
   return (
     <div>
       <div className="alert alert-warning d-flex align-items-center">
         <div>
-          {fetching ? (
-            <span>
-              We&apos;re generating the information schema for this datasource.
-              This may take a minute, depending on the size of the datasource.
-            </span>
-          ) : (
-            <span>{error ? error : informationSchema.error.message}</span>
-          )}
+          <span>{error ? error : informationSchema.error.message}</span>
         </div>
         <button
-          disabled={!!error || fetching}
           className="btn btn-link"
           onClick={async (e) => {
             e.preventDefault();
             onClick();
           }}
         >
-          {fetching && <LoadingSpinner />} Retry
+          Retry
         </button>
       </div>
     </div>
